Move math.hpp to Platforms/CPU
