+++ src/QMCHamiltonians/CoulombPBCAA_CUDA.cpp
@@ -17,7 +17,7 @@
 #include "CoulombPBCAA_CUDA.h"
 #include "Particle/MCWalkerConfiguration.h"
 #include "QMCDrivers/WalkerProperties.h"
-#include "config/stdlib/math.hpp"
+#include "CPU/math.hpp"
 
 namespace qmcplusplus
 {
