+++ src/QMCHamiltonians/CoulombPBCAB_CUDA.cpp
@@ -18,7 +18,7 @@
 #include "CoulombPBCAB_CUDA.h"
 #include "Particle/MCWalkerConfiguration.h"
 #include "QMCDrivers/WalkerProperties.h"
-#include "config/stdlib/math.hpp"
+#include "CPU/math.hpp"
 
 namespace qmcplusplus
 {
