+++ src/QMCWaveFunctions/EinsplineSetBuilderOld.cpp
@@ -24,7 +24,7 @@
 #include "Numerics/HDFSTLAttrib.h"
 #include "OhmmsData/HDFStringAttrib.h"
 #include "ParticleBase/RandomSeqGenerator.h"
-#include "config/stdlib/math.hpp"
+#include "CPU/math.hpp"
 
 namespace qmcplusplus
 {
