+++ src/QMCWaveFunctions/Jastrow/kSpaceJastrow.cpp
@@ -17,7 +17,7 @@
 
 #include "kSpaceJastrow.h"
 #include "LongRange/StructFact.h"
-#include "config/stdlib/math.hpp"
+#include "CPU/math.hpp"
 #include "CPU/e2iphi.h"
 #include <sstream>
 #include <algorithm>
