+++ src/QMCWaveFunctions/LCAO/LCAOrbitalBuilder.cpp
@@ -33,7 +33,7 @@
 #include "hdf/hdf_archive.h"
 #include "Message/CommOperators.h"
 #include "Utilities/ProgressReportEngine.h"
-#include "config/stdlib/math.hpp"
+#include "CPU/math.hpp"
 
 namespace qmcplusplus
 {
